requirements.txt
beautifulsoup4
PyMySQL
requests
urllib3
Pillow
